Hooks module is separated out
